+++ solstice/src/main/java/dev/equo/ide/Catalog.java
@@ -34,6 +34,29 @@ public class Catalog implements Comparable<Catalog> {
 					"3.1.6.v20220511-1359",
 					List.of("org.eclipse.buildship.feature.group"),
 					JDT);
+
+	/** Pure transitive of m2e and others */
+	private static final Catalog LSP4J =
+			new Catalog(
+					"lsp4j", "https://download.eclipse.org/lsp4j/updates/releases/" + V, "0.20.0", List.of());
+
+	/** Pure transitive of m2e and others */
+	private static final Catalog WST =
+			new Catalog(
+					"wst",
+					"https://download.eclipse.org/webtools/downloads/drops/" + V + "/repository/",
+					"R3.28.0/R-3.28.0-20221120050827",
+					List.of());
+
+	public static final Catalog M2E =
+			new Catalog(
+					"m2e",
+					"https://download.eclipse.org/technology/m2e/releases/" + V,
+					"2.1.2",
+					List.of("org.eclipse.m2e.feature.feature.group"),
+					JDT,
+					WST,
+					LSP4J);
 	public static final Catalog PDE =
 			new Catalog("pde", PLATFORM, List.of("org.eclipse.releng.pde.categoryIU"), JDT);
 
@@ -41,7 +64,8 @@ public class Catalog implements Comparable<Catalog> {
 			new Catalog(
 					"kotlin",
 					"https://files.pkg.jetbrains.space/kotlin/p/kotlin-eclipse/main/" + V,
-					"0.8.21", // 0.8.24 is broken in many ways
+					"0.8.21", // 0.8.24 is broken in many
+					// ways
 					List.of(
 							"org.jetbrains.kotlin.feature.feature.group",
 							"org.jetbrains.kotlin.gradle.feature.feature.group"),